requirements.txt
altgraph==0.17.2
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
django-rest-framework==0.1.0
django-safedelete==1.2.2
djangorestframework==3.13.1
future==0.18.2
idna==3.3
pefile==2022.5.30
pycodestyle==2.8.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.28.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
